Introduce the concept of a table
The player will group together their cards an lay them on their table.
This change implements the rules of fives in that each round as a rule
for the amount of cards you must lay before being able to lay other
groups.

When the round ends needs work as currently it doesn't move onto the
next round.

Also the cards on the table need to be drop targets for the cards in the
hand.

You also need to be able to see the cards that the other players have
laid and be able to add cards to them (once you've made your initial lay)